refactor(searcher): Move memory enhancement methods to AdvancedSearcher
Problem:
- enhance_memories_with_query and recall_for_missing_memories were in MemoryPostProcessor
- This caused SearchHandler to depend on Scheduler (cross-layer coupling)
- These are general search enhancement features, not scheduler-specific

Solution:
- Move both methods from MemoryPostProcessor to AdvancedSearcher
- Update SingleCubeView to use self.searcher instead of self.mem_scheduler.post_processor
- Simplify MemoryPostProcessor to only handle filtering and reranking

Benefits:
- Better separation of concerns (search enhancement belongs to Searcher)
- Removes cross-layer dependency (API layer no longer depends on Scheduler)
- Enables reuse (both Scheduler and SearchHandler share same implementation)
- Improves testability (AdvancedSearcher can be tested independently)

Changes:
- advanced_searcher.py: +291 lines (added enhancement methods)
- post_processor.py: -320 lines (removed enhancement, kept filter/rerank)
- single_cube.py: 6 changes (updated call sites)
- test_post_processor.py: -57 lines (removed obsolete tests)

Type safety: Verified via 'AdvancedSearcher as Searcher' alias in tree.py
