+++ src/memos/memories/textual/tree_text_memory/retrieve/advanced_searcher.py
@@ -1,12 +1,15 @@
 import copy
 import time
 
+from concurrent.futures import as_completed
 from typing import Any
 
+from memos.context.context import ContextThreadPoolExecutor
 from memos.embedders.factory import OllamaEmbedder
 from memos.graph_dbs.factory import Neo4jGraphDB
 from memos.llms.factory import AzureLLM, OllamaLLM, OpenAILLM
 from memos.log import get_logger
+from memos.mem_scheduler.utils.misc_utils import extract_json_obj, extract_list_items_in_answer
 from memos.memories.textual.item import TextualMemoryItem, TextualMemoryMetadata
 from memos.memories.textual.tree_text_memory.retrieve.bm25_util import EnhancedBM25
 from memos.memories.textual.tree_text_memory.retrieve.retrieve_utils import (
@@ -16,7 +19,7 @@ from memos.memories.textual.tree_text_memory.retrieve.retrieve_utils import (
 from memos.memories.textual.tree_text_memory.retrieve.searcher import Searcher
 from memos.reranker.base import BaseReranker
 from memos.templates.advanced_search_prompts import PROMPT_MAPPING
-from memos.types.general_types import SearchMode
+from memos.types.general_types import FINE_STRATEGY, FineStrategy, SearchMode
 
 
 logger = get_logger(__name__)
@@ -362,3 +365,289 @@ class AdvancedSearcher(Searcher):
                 continue
         logger.error("Deep search failed, returning original memories")
         return memories
+
+    def enhance_memories_with_query(
+        self,
+        query_history: list[str],
+        memories: list[TextualMemoryItem],
+        batch_size: int | None = None,
+        retries: int = 2,
+    ) -> tuple[list[TextualMemoryItem], bool]:
+        """
+        Enhance memories by adding context and making connections to queries.
+        
+        This method uses LLM to rewrite or recreate memories to better align
+        with the given query history, making them more relevant and contextual.
+        
+        Args:
+            query_history: List of user queries in chronological order
+            memories: List of memory items to enhance
+            batch_size: Optional batch size for parallel processing
+            retries: Number of retries for LLM calls
+            
+        Returns:
+            Tuple of (enhanced_memories, success_flag)
+            - enhanced_memories: Enhanced memory items
+            - success_flag: True if all batches processed successfully
+        """
+        if not memories:
+            logger.warning("[Enhance] ⚠️ skipped (no memories to process)")
+            return memories, True
+
+        num_of_memories = len(memories)
+
+        try:
+            # Single batch path (no parallelization)
+            if batch_size is None or num_of_memories <= batch_size:
+                enhanced_memories, success_flag = self._process_enhancement_batch(
+                    batch_index=0,
+                    query_history=query_history,
+                    memories=memories,
+                    retries=retries,
+                )
+                all_success = success_flag
+            else:
+                # Parallel batch processing
+                batches = self._split_batches(memories=memories, batch_size=batch_size)
+                all_success = True
+                failed_batches = 0
+                
+                with ContextThreadPoolExecutor(max_workers=len(batches)) as executor:
+                    future_map = {
+                        executor.submit(
+                            self._process_enhancement_batch, bi, query_history, texts, retries
+                        ): (bi, s, e)
+                        for bi, (s, e, texts) in enumerate(batches)
+                    }
+                    
+                    enhanced_memories = []
+                    for fut in as_completed(future_map):
+                        bi, s, e = future_map[fut]
+                        batch_memories, ok = fut.result()
+                        enhanced_memories.extend(batch_memories)
+                        
+                        if not ok:
+                            all_success = False
+                            failed_batches += 1
+                            
+                logger.info(
+                    f"[Enhance] ✅ multi-batch done | batches={len(batches)} | "
+                    f"enhanced={len(enhanced_memories)} | failed_batches={failed_batches} | "
+                    f"success={all_success}"
+                )
+
+        except Exception as e:
+            logger.error(f"[Enhance] ❌ fatal error: {e}", exc_info=True)
+            all_success = False
+            enhanced_memories = memories
+
+        if len(enhanced_memories) == 0:
+            enhanced_memories = []
+            logger.error("[Enhance] ❌ fatal error: enhanced_memories is empty", exc_info=True)
+            
+        return enhanced_memories, all_success
+
+    def _process_enhancement_batch(
+        self,
+        batch_index: int,
+        query_history: list[str],
+        memories: list[TextualMemoryItem],
+        retries: int,
+    ) -> tuple[list[TextualMemoryItem], bool]:
+        """
+        Process a single batch of memories for enhancement.
+        
+        This method handles retry logic and strategy-specific enhancement
+        (REWRITE vs RECREATE).
+        """
+        attempt = 0
+        text_memories = [one.memory for one in memories]
+
+        prompt = self._build_enhancement_prompt(
+            query_history=query_history, batch_texts=text_memories
+        )
+
+        llm_response = None
+        while attempt <= max(0, retries) + 1:
+            try:
+                llm_response = self.process_llm.generate([{"role": "user", "content": prompt}])
+                processed_text_memories = extract_list_items_in_answer(llm_response)
+                
+                if len(processed_text_memories) > 0:
+                    enhanced_memories = self._create_enhanced_memories(
+                        processed_text_memories=processed_text_memories,
+                        original_memories=memories,
+                    )
+                    
+                    logger.info(
+                        f"[enhance_memories_with_query] ✅ done | Strategy={FINE_STRATEGY} | "
+                        f"batch={batch_index}"
+                    )
+                    return enhanced_memories, True
+                else:
+                    raise ValueError(
+                        f"Fail to run memory enhancement; retry {attempt}/{max(1, retries) + 1}; "
+                        f"processed_text_memories: {processed_text_memories}"
+                    )
+                    
+            except Exception as e:
+                attempt += 1
+                time.sleep(1)
+                logger.debug(
+                    f"[enhance_memories_with_query][batch={batch_index}] "
+                    f"🔁 retry {attempt}/{max(1, retries) + 1} failed: {e}"
+                )
+                
+        logger.error(
+            f"Fail to run memory enhancement; prompt: {prompt};\n llm_response: {llm_response}",
+            exc_info=True,
+        )
+        return memories, False
+
+    def _build_enhancement_prompt(
+        self, query_history: list[str], batch_texts: list[str]
+    ) -> str:
+        """Build the LLM prompt for memory enhancement."""
+        if len(query_history) == 1:
+            query_history_formatted = query_history[0]
+        else:
+            query_history_formatted = (
+                [f"[{i}] {query}" for i, query in enumerate(query_history)]
+                if len(query_history) > 1
+                else query_history[0]
+            )
+
+        # Include numbering for rewrite mode to help LLM reference original memory IDs
+        if FINE_STRATEGY == FineStrategy.REWRITE:
+            text_memories = "\n".join([f"- [{i}] {mem}" for i, mem in enumerate(batch_texts)])
+            prompt_name = "memory_rewrite_enhancement"
+        else:
+            text_memories = "\n".join([f"- {mem}" for i, mem in enumerate(batch_texts)])
+            prompt_name = "memory_recreate_enhancement"
+            
+        return self.build_prompt(
+            prompt_name,
+            query_history=query_history_formatted,
+            memories=text_memories,
+        )
+
+    def _create_enhanced_memories(
+        self,
+        processed_text_memories: list[str],
+        original_memories: list[TextualMemoryItem],
+    ) -> list[TextualMemoryItem]:
+        """
+        Create enhanced memory items based on the processing strategy.
+        
+        Supports two strategies:
+        - RECREATE: Create new memory items with enhanced text
+        - REWRITE: Rewrite existing memories while preserving metadata
+        """
+        enhanced_memories = []
+        user_id = original_memories[0].metadata.user_id
+
+        if FINE_STRATEGY == FineStrategy.RECREATE:
+            for new_mem in processed_text_memories:
+                enhanced_memories.append(
+                    TextualMemoryItem(
+                        memory=new_mem,
+                        metadata=TextualMemoryMetadata(
+                            user_id=user_id, memory_type="LongTermMemory"
+                        ),
+                    )
+                )
+                
+        elif FINE_STRATEGY == FineStrategy.REWRITE:
+            # Parse index from each processed line and rewrite corresponding original memory
+            def _parse_index_and_text(s: str) -> tuple[int | None, str]:
+                import re
+
+                s = (s or "").strip()
+                # Preferred: [index] text
+                m = re.match(r"^\s*\[(\d+)\]\s*(.+)$", s)
+                if m:
+                    return int(m.group(1)), m.group(2).strip()
+                # Fallback: index: text or index - text
+                m = re.match(r"^\s*(\d+)\s*[:\-\)]\s*(.+)$", s)
+                if m:
+                    return int(m.group(1)), m.group(2).strip()
+                return None, s
+
+            idx_to_original = dict(enumerate(original_memories))
+            for j, item in enumerate(processed_text_memories):
+                idx, new_text = _parse_index_and_text(item)
+                if idx is not None and idx in idx_to_original:
+                    orig = idx_to_original[idx]
+                else:
+                    # Fallback: align by order if index missing/invalid
+                    orig = original_memories[j] if j < len(original_memories) else None
+                    
+                if not orig:
+                    continue
+                    
+                enhanced_memories.append(
+                    TextualMemoryItem(
+                        id=orig.id,
+                        memory=new_text,
+                        metadata=orig.metadata,
+                    )
+                )
+        else:
+            logger.error(f"Fine search strategy {FINE_STRATEGY} not exists")
+
+        return enhanced_memories
+
+    @staticmethod
+    def _split_batches(
+        memories: list[TextualMemoryItem], batch_size: int
+    ) -> list[tuple[int, int, list[TextualMemoryItem]]]:
+        """Split memories into batches for parallel processing."""
+        batches: list[tuple[int, int, list[TextualMemoryItem]]] = []
+        start = 0
+        n = len(memories)
+        while start < n:
+            end = min(start + batch_size, n)
+            batches.append((start, end, memories[start:end]))
+            start = end
+        return batches
+
+    def recall_for_missing_memories(
+        self,
+        query: str,
+        memories: list[str],
+    ) -> tuple[str, bool]:
+        """
+        Analyze memories and generate hint for additional recall.
+        
+        This method uses LLM to determine if the current memories are sufficient
+        or if additional recall is needed, along with a hint for the recall query.
+        
+        Args:
+            query: Original user query
+            memories: List of currently retrieved memory texts
+            
+        Returns:
+            Tuple of (hint, trigger_recall)
+            - hint: Suggested query for additional recall
+            - trigger_recall: Whether to trigger additional recall
+        """
+        text_memories = "\n".join([f"- {mem}" for i, mem in enumerate(memories)])
+
+        prompt = self.build_prompt(
+            template_name="enlarge_recall",
+            query=query,
+            memories_inline=text_memories,
+        )
+        llm_response = self.process_llm.generate([{"role": "user", "content": prompt}])
+
+        json_result: dict = extract_json_obj(llm_response)
+
+        logger.info(
+            f"[recall_for_missing_memories] ✅ done | prompt={prompt} | "
+            f"llm_response={llm_response}"
+        )
+
+        hint = json_result.get("hint", "")
+        if len(hint) == 0:
+            return hint, False
+        return hint, json_result.get("trigger_recall", False)
